ai-agent/requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain
langchain-openai
tavily-python
python-multipart
pydantic-settings
python-dotenv
httpx
pymupdf4llm
langsmith
langchain-groq
arxiv
pytest